getMany() with <entity>/ids/1,2,3 with an OpenAPI compliant array<int64> parameter:
OpenAPI spec compliant means that an OpenAPI client sends the right
request (and at least IntelliJ does that).

comma-separate is also easy to read, but obviously means that
identifiers cannot have comma.

A disadvantage is that the Chrome Developer tools shows such a request
as just "1,2,3", so you do not immediately see the type.